codex(character): save-v3 [4] TM fix
클라 LEVEL_CAP 200→1000 서버 parity, async 콜백 TWeakObjectPtr 가드,
adopt 시 level clamp + payload roundtrip 테스트 강화, Quest/Pet/Season
auth/characterId 가드. 서버 characters.total_exp→bigint(level 1000 int32
초과) + migration 0006 + 회귀 테스트. Codex 커밋 누락 → PM 보완.

+++ server/src/modules/save/save.repo.ts
@@ -42,7 +42,7 @@ export class SaveRepoPg {
          set level = greatest(level, $2),
              rebirth_count = greatest(rebirth_count, $3),
              gold = greatest(gold, coalesce(($4::jsonb ->> 'gold')::int, gold)),
-             total_exp = greatest(total_exp, coalesce(($4::jsonb ->> 'totalExp')::int, total_exp)),
+             total_exp = greatest(total_exp, coalesce(($4::jsonb ->> 'totalExp')::bigint, total_exp)),
              last_seen_at = coalesce(to_timestamp(($4::jsonb ->> 'lastSeenUnixSec')::double precision), last_seen_at, now()),
              last_save_at = now()
          where id = $1`,
